clean_version/requirements.txt
# Core Flask Application
Flask==3.0.0
Werkzeug==3.0.0

# Database & Vector Search
supabase==2.0.0
psycopg2-binary==2.9.7

# AI Providers
openai==1.35.0
anthropic==0.25.0
google-generativeai==0.5.0

# Video Processing  
yt-dlp==2024.7.7
ffmpeg-python==0.2.0

# HTTP Requests
requests==2.31.0
httpx==0.24.1

# Environment & Configuration
python-dotenv==1.0.0

# Task Scheduling
APScheduler==3.10.4

# Audio Processing (optional)
gTTS==2.3.2

# Development
python-dateutil==2.8.2